pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "DocFusionX Server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.0",
    "fastapi[standard]>=0.128.5",
    "mistralai>=1.12.0",
    "uvicorn>=0.40.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "docling>=2.72.0",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.2.2",
    "langchain>=1.2.10",
    "rank-bm25>=0.2.2",
    "langchain-text-splitters>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-chroma>=1.1.0",
]
